getNextMove -> made it a recursive call
The algorithm was still set up and partly handled in NextMove.main so moved it to NextMove so that it can be executed from main.main for future iterations and also from NextMove.main to satisfy HackerRank. The algorithm now as written detects the score of the grid and adjusts the parameters to search vertically and horizontally to keep the timeout low but maximize depth.  Hopefully, future iterations of the project will give the user the ability to choose the parameters.